+++ examples/complete/locals.tf
@@ -1,0 +1,4 @@
+locals {
+  admin    = github_team.admin.id
+  maintain = github_team.maintain.id
+}