+++ main.tf
@@ -41,11 +41,13 @@ resource "github_repository" "main" {
 }
 
 resource "github_branch" "branch" {
-  count      = length(var.branch) == 0 ? 0 : length(var.branch)
-  repository = github_repository.main.name
-  branch     = element(var.branch, count.index)
+  count         = length(var.branch) == 0 ? 0 : length(var.branch)
+  repository    = github_repository.main.name
+  branch        = element(var.branch, count.index)
+  source_branch = var.default_branch
   depends_on = [
-    github_repository.main
+    github_repository.main,
+    github_branch.default
   ]
 }
 
@@ -129,8 +131,8 @@ resource "github_branch_protection" "main" {
   dynamic "required_status_checks" {
     for_each = local.required_status_checks
     content {
-      strict   = each.key
-      contexts = [each.value]
+      strict   = required_status_checks.value.strict
+      contexts = required_status_checks.value.contexts
     }
   }
 
@@ -155,4 +157,14 @@ resource "github_issue_label" "main" {
   name        = each.key
   color       = each.value.color
   description = each.value.description
+}
+
+resource "github_actions_secret" "main" {
+  for_each        = var.secrets
+  repository      = github_repository.main.name
+  secret_name     = each.key
+  encrypted_value = data.sodium_encrypted_item.main[each.key].encrypted_value_base64
+  depends_on = [
+    github_repository.main
+  ]
 }