+++ CHANGELOG.md
@@ -10,9 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix: CKV_GIT_6: "Ensure all commits GPG signed"
 - fix: CKV_GIT_5: "Ensure at least two approving reviews for PRs"
 
+## [1.3.0] - 2022-11-21
+### Changes
+- feat: added github actions secret resource to create and manage repository secrets
+- fix: source branch for additional branches being created
+- fix: required status checks arguments
+
 ## [1.2.0] - 2022-08-30
 ### Changes
-- feat: Resouce to create issues labels
+- feat: Resource to create issues labels
 
 ## [1.1.0] - 2022-05-18
 The following  additions, changes and fixes were made in this release.
@@ -41,7 +47,8 @@ Here following resources were added in this release.
 - Github repository resource
 - Initial commit
 
-[Unreleased]: https://github.com/boldlink/terraform-github-repository/1.2.0...HEAD
-[1.1.2]: https://github.com/boldlink/terraform-github-repository/releases/tag/1.2.0
+[Unreleased]: https://github.com/boldlink/terraform-github-repository/1.3.0...HEAD
+[1.3.0]: https://github.com/boldlink/terraform-github-repository/releases/tag/1.3.0
+[1.2.0]: https://github.com/boldlink/terraform-github-repository/releases/tag/1.2.0
 [1.1.0]: https://github.com/boldlink/terraform-github-repository/releases/tag/1.1.0
 [1.0.0]: https://github.com/boldlink/terraform-github-repository/releases/tag/1.0.0